final2

+++ main.py
@@ -1,10 +1,9 @@
 import tkinter as tk
-from tkinter import messagebox, simpledialog
+from tkinter import messagebox, simpledialog, filedialog
 import mysql.connector
 from mysql.connector import Error
 from openpyxl import Workbook
-from openpyxl.styles import Font
-from datetime import datetime, time
+from datetime import datetime, time, timedelta
 import unicodedata
 
 # Função para conectar ao banco de dados
@@ -18,8 +17,6 @@ def conectar_bd():
         )
         if conn.is_connected():
             print('Conexão ao banco de dados bem-sucedida.')
-            # Adicionando commit aqui
-            conn.commit()
             return conn
     except Error as e:
         print(f"Erro ao conectar ao banco de dados: {e}")
@@ -29,6 +26,60 @@ def remover_acentos(texto):
     # Usando NFKD para decompor e filtrando apenas caracteres ASCII
     return ''.join(c for c in unicodedata.normalize('NFKD', texto) if not unicodedata.combining(c))
 
+# Variável global para armazenar o tipo de ponto selecionado
+tipo_ponto_selecionado = None
+
+# Função para registrar ponto através do teclado numérico
+def registrar_ponto_teclado(event):
+    global tipo_ponto_selecionado
+    if event.char == '1':
+        tipo_ponto_selecionado = "Entrada Manhã"
+    elif event.char == '2':
+        tipo_ponto_selecionado = "Saída Manhã"
+    elif event.char == '3':
+        tipo_ponto_selecionado = "Entrada Tarde"
+    elif event.char == '4':
+        tipo_ponto_selecionado = "Saída Tarde"
+    else:
+        return
+    registrar_ponto(tipo_ponto_selecionado)
+
+# Função para registrar ponto
+def registrar_ponto(tipo_ponto):
+    # Criar uma janela para entrada de código de funcionário
+    def abrir_janela_registro():
+        codigo_funcionario = simpledialog.askinteger("Registrar Ponto", "Digite o código do funcionário:")
+        if codigo_funcionario:
+            verificar_registro(codigo_funcionario, tipo_ponto)
+
+    # Função para verificar o registro de ponto
+    def verificar_registro(codigo_funcionario, tipo_ponto):
+        try:
+            conn = conectar_bd()
+            if conn:
+                cursor = conn.cursor()
+                cursor.execute("SELECT nome FROM Funcionarios WHERE codigo = %s", (codigo_funcionario,))
+                nome_funcionario = cursor.fetchone()
+                if nome_funcionario:
+                    # Pop-up de confirmação
+                    confirmacao = messagebox.askokcancel("Confirmação", f"Você deseja registrar um ponto de {tipo_ponto} para {nome_funcionario[0]}?")
+                    if confirmacao:
+                        # Registrar o ponto
+                        cursor.execute("INSERT INTO RegistrosPonto (codigo_funcionario, tipo_ponto, data_hora) VALUES (%s, %s, NOW())", (codigo_funcionario, tipo_ponto))
+                        conn.commit()
+                        messagebox.showinfo("Sucesso", f"Ponto {tipo_ponto} registrado com sucesso para {nome_funcionario[0]}.")
+                else:
+                    messagebox.showerror("Erro", "Código de funcionário não encontrado.")
+        except Error as e:
+            messagebox.showerror("Erro", f"Erro ao registrar ponto: {e}")
+        finally:
+            if 'cursor' in locals() and cursor:
+                cursor.close()
+            if 'conn' in locals() and conn:
+                conn.close()
+
+    # Chamar a função para abrir a janela de registro
+    abrir_janela_registro()
 
 # Função para editar registro de ponto
 def editar_registro():
@@ -116,60 +167,6 @@ def excluir_registro():
     # Chamar a função para abrir a janela de exclusão
     abrir_janela_exclusao()
 
-    # Chamar a função para abrir a janela de exclusão
-    abrir_janela_exclusao()
-
-# Função para registrar ponto
-def registrar_ponto(tipo_ponto):
-    # Criar uma janela para entrada de código de funcionário
-    def abrir_janela_registro():
-        codigo_funcionario = simpledialog.askinteger("Registrar Ponto", "Digite o código do funcionário:")
-        if codigo_funcionario:
-            verificar_registro(codigo_funcionario, tipo_ponto)
-
-    # Função para verificar o registro de ponto
-    def verificar_registro(codigo_funcionario, tipo_ponto):
-        try:
-            conn = conectar_bd()
-            if conn:
-                cursor = conn.cursor()
-                cursor.execute("SELECT nome FROM Funcionarios WHERE codigo = %s", (codigo_funcionario,))
-                nome_funcionario = cursor.fetchone()
-                if nome_funcionario:
-                    # Verificar se o funcionário já registrou entrada manhã
-                    if tipo_ponto == "Saída Manhã":
-                        cursor.execute("SELECT tipo_ponto FROM RegistrosPonto WHERE codigo_funcionario = %s AND tipo_ponto = 'Entrada Manhã'", (codigo_funcionario,))
-                        entrada_manha = cursor.fetchone()
-                        if not entrada_manha:
-                            messagebox.showerror("Erro", "Você não registrou entrada pela manhã.")
-                            return
-                    # Verificar se o funcionário já registrou entrada tarde
-                    if tipo_ponto == "Saída Tarde":
-                        cursor.execute("SELECT tipo_ponto FROM RegistrosPonto WHERE codigo_funcionario = %s AND tipo_ponto IN ('Entrada Manhã', 'Entrada Tarde')", (codigo_funcionario,))
-                        entrada_tarde = cursor.fetchone()
-                        if not entrada_tarde:
-                            messagebox.showerror("Erro", "Você não registrou entrada.")
-                            return
-                    # Pop-up de confirmação
-                    confirmacao = messagebox.askokcancel("Confirmação", f"Você deseja registrar um ponto de {tipo_ponto} para {nome_funcionario[0]}?")
-                    if confirmacao:
-                        # Registrar o ponto
-                        cursor.execute("INSERT INTO RegistrosPonto (codigo_funcionario, tipo_ponto, data_hora) VALUES (%s, %s, NOW())", (codigo_funcionario, tipo_ponto))
-                        conn.commit()
-                        messagebox.showinfo("Sucesso", f"Ponto {tipo_ponto} registrado com sucesso para {nome_funcionario[0]}.")
-                else:
-                    messagebox.showerror("Erro", "Código de funcionário não encontrado.")
-        except Error as e:
-            messagebox.showerror("Erro", f"Erro ao registrar ponto: {e}")
-        finally:
-            if 'cursor' in locals() and cursor:
-                cursor.close()
-            if 'conn' in locals() and conn:
-                conn.close()
-
-    # Chamar a função para abrir a janela de registro
-    abrir_janela_registro()
-
 # Função para fazer login
 def fazer_login():
     # Cria uma janela de login
@@ -211,23 +208,28 @@ def gerar_relatorio():
                 cursor.execute("SELECT f.codigo, f.nome, DATE_FORMAT(rp.data_hora, '%d/%m/%Y %H:%i:%s'), rp.tipo_ponto FROM Funcionarios f INNER JOIN RegistrosPonto rp ON f.codigo = rp.codigo_funcionario ORDER BY f.codigo, rp.data_hora")
                 relatorio = cursor.fetchall()
 
-                # Criar arquivo Excel
-                wb = Workbook()
-                for linha in relatorio:
-                    codigo_funcionario, nome_funcionario, data_hora, tipo_ponto = linha
-                    nome_planilha = remover_acentos(nome_funcionario)
-                    if nome_planilha not in wb.sheetnames:
-                        ws = wb.create_sheet(title=nome_planilha)
-                        ws.append(["Código", "Nome", "Data/Hora", "Tipo de Ponto"])
-                    ws.append([codigo_funcionario, nome_funcionario, data_hora, tipo_ponto])
-                wb.remove(wb["Sheet"])  # Remover a planilha padrão criada automaticamente
-                wb.save("relatorio_ponto.xlsx")
+                # Solicitar ao usuário o local para salvar o relatório
+                file_path = filedialog.asksaveasfilename(defaultextension=".xlsx", filetypes=[("Arquivo Excel", "*.xlsx")])
+                if file_path:
+                    # Criar arquivo Excel
+                    wb = Workbook()
+                    for linha in relatorio:
+                        codigo_funcionario, nome_funcionario, data_hora, tipo_ponto = linha
+                        nome_planilha = remover_acentos(nome_funcionario)
+                        if nome_planilha not in wb.sheetnames:
+                            ws = wb.create_sheet(title=nome_planilha)
+                            ws.append(["Código", "Nome", "Data/Hora", "Tipo de Ponto"])
+                        
+                        ws.append([codigo_funcionario, nome_funcionario, data_hora, tipo_ponto])
 
-                # Limpar apenas os dados de registros do banco de dados
-                cursor.execute("DELETE FROM RegistrosPonto")
-                conn.commit()
+                    wb.remove(wb["Sheet"])  # Remover a planilha padrão criada automaticamente
+                    wb.save(file_path)
 
-                messagebox.showinfo("Relatório", "Relatório gerado com sucesso.")
+                    # Limpar apenas os dados de registros do banco de dados
+                    cursor.execute("DELETE FROM RegistrosPonto")
+                    conn.commit()
+
+                    messagebox.showinfo("Relatório", "Relatório gerado com sucesso.")
         except Error as e:
             messagebox.showerror("Erro", f"Erro ao gerar relatório: {e}")
         finally:
@@ -236,6 +238,24 @@ def gerar_relatorio():
             if 'conn' in locals() and conn:
                 conn.close()
 
+# Função para verificar se o horário está dentro do intervalo permitido
+def verificar_horario_permitido(data_hora, tipo_ponto):
+    horario = datetime.strptime(data_hora, '%d/%m/%Y %H:%M:%S').time()
+    dia_semana = datetime.strptime(data_hora, '%d/%m/%Y %H:%M:%S').weekday()
+
+    if tipo_ponto == "Entrada Manhã" or tipo_ponto == "Saída Manhã":
+        if horario >= time(7, 15) and horario <= time(11, 30) and dia_semana < 5:
+            return True
+        else:
+            return False
+    elif tipo_ponto == "Entrada Tarde" or tipo_ponto == "Saída Tarde":
+        if horario >= time(13, 0) and horario <= time(16, 45) and dia_semana < 5:
+            return True
+        else:
+            return False
+    else:
+        return False
+    
 # Função principal para criar a interface gráfica
 def criar_interface():
     root = tk.Tk()
@@ -256,18 +276,21 @@ def criar_interface():
 
     # Quadrado para registros de horários
     frame_registros = tk.Frame(root, borderwidth=2, relief="groove")
-    frame_registros.place(relx=0.15, rely=0.3, anchor=tk.CENTER)
+    frame_registros.place(relx=0.15, rely=0.3, anchor=tk.CENTER, )
 
-    button_entrada_manha = tk.Button(frame_registros, text="Registrar Entrada Manhã", command=lambda: registrar_ponto("Entrada Manhã"))
+    # Adicione um evento <Key> para a janela principal
+    root.bind("<Key>", registrar_ponto_teclado)
+
+    button_entrada_manha = tk.Button(frame_registros, text="Registrar Entrada Manhã\n(1)", command=lambda: registrar_ponto("Entrada Manhã"))
     button_entrada_manha.grid(row=0, column=0, padx=10, pady=5)
 
-    button_saida_manha = tk.Button(frame_registros, text="Registrar Saída Manhã", command=lambda: registrar_ponto("Saída Manhã"))
+    button_saida_manha = tk.Button(frame_registros, text="Registrar Saída Manhã\n(2)", command=lambda: registrar_ponto("Saída Manhã"))
     button_saida_manha.grid(row=1, column=0, padx=10, pady=5)
 
-    button_entrada_tarde = tk.Button(frame_registros, text="Registrar Entrada Tarde", command=lambda: registrar_ponto("Entrada Tarde"))
+    button_entrada_tarde = tk.Button(frame_registros, text="Registrar Entrada Tarde\n(3)", command=lambda: registrar_ponto("Entrada Tarde"))
     button_entrada_tarde.grid(row=2, column=0, padx=10, pady=5)
 
-    button_saida_tarde = tk.Button(frame_registros, text="Registrar Saída Tarde", command=lambda: registrar_ponto("Saída Tarde"))
+    button_saida_tarde = tk.Button(frame_registros, text="Registrar Saída Tarde\n(4)", command=lambda: registrar_ponto("Saída Tarde"))
     button_saida_tarde.grid(row=3, column=0, padx=10, pady=5)
 
     # Quadrado para excluir/editar registros
